+++ source/extensions/lightspeed.ui_scene.light_manipulator/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.7]
+## Fixed
+- Fixed crash when stage is not available
+
 ## [1.0.6]
 ## Changed
 - Update to Kit 106.5